% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_table.R
\name{likert_table}
\alias{likert_table}
\title{Create table from likert function}
\usage{
likert_table(x, include_ns = TRUE, cross = NULL)
}
\arguments{
\item{x}{A dataframe with only the variables you want in this function call.}

\item{include_ns}{Should the table include N's.}

\item{cross}{What variable to cross table with.}
}
\value{
Likert result dataframe
}
\description{
Create table from likert function
}
