% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_table_style.R
\name{likert_table_style}
\alias{likert_table_style}
\title{Add gt style to table}
\usage{
likert_table_style(x, table_title, column_map = NULL)
}
\arguments{
\item{x}{A dataframe}

\item{table_title}{Title for gt Table}

\item{column_map}{dataframe with variable labels in 2 columns named Item and Label}
}
\value{
Styled gt table
}
\description{
Add gt style to table
}
