% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_tidy_table.R
\name{likert_tidy_table}
\alias{likert_tidy_table}
\title{Create table from likert function}
\usage{
likert_tidy_table(question)
}
\arguments{
\item{question}{question number}
}
\value{
Likert result dataframe
}
\description{
Create table from likert function
}
