% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_tidy_table_style.R
\name{likert_tidy_table_style}
\alias{likert_tidy_table_style}
\title{Create styled gt likert table}
\usage{
likert_tidy_table_style(x, table_title)
}
\arguments{
\item{x}{A dataframe}

\item{table_title}{Title for gt Table}
}
\value{
Styled gt table
}
\description{
Create styled gt likert table
}
