% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_response_table.R,
%   R/multi_response_table_style.R
\name{multi_response_table}
\alias{multi_response_table}
\alias{multi_response_table_style}
\title{Function for multiple-response question frequency table}
\usage{
multi_response_table(data, ..., cross = NULL)

multi_response_table_style(x, table_title)
}
\arguments{
\item{data}{Survey data with multireponse split into multiple columns}

\item{...}{selection of variable using expss selection helpers}

\item{cross}{column variable}

\item{x}{result of multi_response_table}

\item{table_title}{The text for the title of the table}
}
\value{
Multi-response dataframe

Styled multi-response dataframe
}
\description{
\code{multi_response_table} creates a dataframe with frequency and percent of cases for multiple-response questions

\code{multi_response_table_style} creates a GT styled multi-response table from the output of \code{multi_response_table}
}
\examples{
# a simple multi-response table
multi_response_table(rchabits, expss::mdset(expss::..p("q5_\\\\d"))) |>
    multi_response_table_style("Types of Cuisine")

# a multi-response table with a cross variable
multi_response_table(rchabits, cross = q1, expss::mdset(expss::..p("q5_\\\\d"))) |>
    multi_response_table_style("Types of Cuisine X Age")


}
