% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_response_table_style.R
\name{multi_response_table_style}
\alias{multi_response_table_style}
\title{Function for styling Multiple-response question frequency table in RC branding}
\usage{
multi_response_table_style(x)
}
\arguments{
\item{x}{result of multi_response_table}
}
\value{
Styled multi-response dataframe
}
\description{
Function for styling Multiple-response question frequency table in RC branding
}
\examples{
\dontrun{multi_response_table(data = survey_data, mdset_t("q1_2_{1:11}")) |> 
multi_response_table_style()
}

}
