% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_report.R
\name{new_report}
\alias{new_report}
\title{Create a new report}
\usage{
new_report(dir, token = Sys.getenv("GIT_ROSOV"))
}
\arguments{
\item{dir}{Directory for the report}

\item{token}{A Github token with read access of the
github.com/rosovconsulting/rosov-report-template repository.
Default to the GIT_ROSOV environment variable.}
}
\description{
Init a new directory for a report based on Rosov
Consulting Typst template defined here:
https://github.com/rosovconsulting/rosov-report-template
}
\examples{
\dontrun{
new_report("example-report")
}

}
