% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_ends_table.R
\name{open_ends_table}
\alias{open_ends_table}
\title{Create searchable, paginated table of open-ended responses}
\usage{
open_ends_table(data, varname)
}
\arguments{
\item{data}{dataframe}

\item{varname}{column name}
}
\value{
searchable, paginated datatable of open-ended responses
}
\description{
Create searchable, paginated table of open-ended responses
}
