% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rchabits}
\alias{rchabits}
\title{Rosov Consulting Habits Dataset}
\format{
\subsection{\code{rchabits}}{

A data frame with 20 rows and 32 columns:
\describe{
\item{response_id}{Qualtrics response ID}
\item{q1}{Age}
\item{q2}{Gender}
\item{q2_4_text}{Gender write-in}
\item{q3_1, q3_5}{Eating habits}
\item{q4_1, q4_5}{Restaurant choice factors}
\item{q5_1, q5_9}{Types of cuisine}
\item{q6_1, q6_8}{Streaming services}
...
}
}
}
\source{
\url{https://rosov.co1.qualtrics.com/survey-builder/SV_cUQAFC3gyrCZgYC/edit}
}
\usage{
rchabits
}
\description{
A generated survey dataset about habits for example purposes
}
\keyword{datasets}
