% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_style.R
\name{table_style}
\alias{table_style}
\title{Add gt style to table}
\usage{
table_style(x, table_title)
}
\arguments{
\item{x}{A dataframe}

\item{table_title}{Title for gt Table}
}
\value{
Styled gt table
}
\description{
Add gt style to table
}
