% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_summary.R
\name{view_summary}
\alias{view_summary}
\title{Create dataframe summary and view in browser}
\usage{
view_summary(data)
}
\arguments{
\item{data}{dataframe to summarize}
}
\value{
Opens a browser window with the dataframe summary
}
\description{
Create dataframe summary and view in browser
}
