% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_column_map.R
\name{create_column_map}
\alias{create_column_map}
\title{Create column map}
\usage{
create_column_map(data)
}
\arguments{
\item{data}{A data frame with variable labels}
}
\value{
A dataframe
}
\description{
Create column map
}
\examples{
\dontrun{col_map <- create_column_map(survey_data)}

}
