% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_table.R, R/likert_table_style.R
\name{likert_table}
\alias{likert_table}
\alias{likert_table_style}
\title{Create table from likert function}
\usage{
likert_table(x, include_ns = TRUE, cross = NULL)

likert_table_style(x, table_title, column_map = NULL)
}
\arguments{
\item{x}{A dataframe}

\item{include_ns}{Should the table include N's.}

\item{cross}{What variable to cross table with.}

\item{table_title}{Title for gt Table}

\item{column_map}{dataframe with variable labels in 2 columns named Item and Label}
}
\value{
Likert result dataframe

Styled gt table
}
\description{
\code{likert_table} creates a dataframe with proportions and total n's for likert matrix questions

\code{likert_table_style} creates a GT styled likert table from the output of \code{likert_table}
}
\examples{
# a simple likert table
rchabits |>
   dplyr::select(tidyselect::starts_with("q3_")) |>
   likert_table()

# a likert table with a cross variable
\dontrun{rchabits |>
    dplyr::select(q1, tidyselect::starts_with("q3_")) |>
    likert_table(cross = q1) |>
    likert_table_style("Title")}


}
