% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_response_table.R
\name{multi_response_table}
\alias{multi_response_table}
\title{Function for Multiple-response question frequency table}
\usage{
multi_response_table(data, ...)
}
\arguments{
\item{data}{Survey data with multireponse split into multiple columns}

\item{...}{expss selection}
}
\value{
Multi-response dataframe
}
\description{
Function for Multiple-response question frequency table
}
\examples{
multi_response_table(data = survey_data, mdset_t("q1_2_{1:11}")) \%>\% 
multi_response_table_style()


}
